pyproject.toml
[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[tool.poetry]
name = "jwty-mcjwt-face"
version = "0.1.0"
description = ""
authors = ["Matt Pladna <pladdypants@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
falcon = "^2.0.0"
PyJWT = {extras = ["crypto"], version = "^2.0.1"}
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
bandit = "^1.7.0"
safety = "^1.10.3"
pytest-cov = "^2.11.1"
codecov = "^2.1.11"
pytest = "^6.2.2"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
